--- packages/components/src/components/primitives/Snackbar/DefaultSnackbar/index.ts
@@ -1,1 +1,0 @@
-export * from './DefaultSnackbar';
