+++ packages/components/src/components/primitives/Snackbar/index.ts
@@ -1,0 +1,1 @@
+export * from './Snackbar';
